+++ src/init.lua
@@ -5,7 +5,7 @@
 	@file Scrypt.lua
     @shared
     @author zblox164
-    @version 0.0.4-alpha
+    @version 0.0.41-alpha
     @since 2024-12-17
 --]]
 
@@ -39,11 +39,17 @@ export type Result<T> = {
     Error: string?
 }
 
-export type Packet = number | string | {ClientPacketData} | boolean | Instance | buffer | Player | CFrame | Vector3 | Vector2 | Color3 | UDim2 | UDim | Enum | BrickColor
+export type Packet = number | string | {Packet} | boolean | Instance | buffer | Player | CFrame | Vector3 | Vector2 | Color3 | UDim2 | UDim | Enum | BrickColor
 export type ClientPacketData = {
     Data: Packet,
     Reliable: boolean
 }
+
+export type ServerPacketData = {
+    Data: Packet,
+    Reliable: boolean
+}
+
 
 export type Service = {
 	["Name"]: string,